Replace the zod-openapi templates with hono-openapi templates
