+++ templates/base-openapi/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "hono-neon-drizzle-template-openapi",
+  "name": "hono-neon-drizzle-template-hono-openapi",
   "scripts": {
     "dev": "wrangler dev src/index.ts",
     "deploy": "wrangler deploy --minify src/index.ts",
@@ -7,21 +7,20 @@
     "db:migrate": "drizzle-kit migrate",
     "db:seed": "tsx seed.ts",
     "db:setup": "npm run db:generate && npm run db:migrate && npm run db:seed",
-    "db:studio": "drizzle-kit studio",
-    "lint": "biome check --write src",
-    "format": "biome format --write src"
+    "db:studio": "drizzle-kit studio"
   },
   "dependencies": {
-    "@fiberplane/hono": "0.5.2",
-    "@hono/zod-openapi": "^0.18.4",
+    "@fiberplane/hono": "0.4.4-beta.2",
+    "@hono/zod-validator": "^0.4.3",
     "@neondatabase/serverless": "^0.10.4",
     "dotenv": "^16.4.7",
     "drizzle-orm": "^0.39.3",
     "hono": "^4.7.0",
-    "zod": "^3.23.8"
+    "hono-openapi": "^0.4.6",
+    "zod": "^3.23.8",
+    "zod-openapi": "^4.2.4"
   },
   "devDependencies": {
-    "@biomejs/biome": "1.9.4",
     "@cloudflare/workers-types": "^4.20250321.0",
     "drizzle-kit": "^0.30.4",
     "drizzle-seed": "^0.3.1",
