+++ templates/base-openapi/src/index.ts
@@ -1,9 +1,13 @@
 import { neon } from "@neondatabase/serverless";
 import { eq } from "drizzle-orm";
 import { drizzle, type NeonHttpDatabase } from "drizzle-orm/neon-http";
-import { OpenAPIHono, createRoute, z } from "@hono/zod-openapi";
+import { describeRoute, openAPISpecs } from "hono-openapi";
+import { Hono } from "hono";
 import * as schema from "./db/schema";
 import { createFiberplane } from "@fiberplane/hono";
+import z from "zod";
+import { resolver, validator as zValidator } from "hono-openapi/zod";
+import "zod-openapi/extend";
 
 // Types for environment variables and context
 type Bindings = {
@@ -16,7 +20,7 @@ type Variables = {
 
 // Create the app with type-safe bindings and variables
 // For more information on OpenAPIHono, see: https://hono.dev/examples/zod-openapi
-const app = new OpenAPIHono<{ Bindings: Bindings; Variables: Variables }>();
+const app = new Hono<{ Bindings: Bindings; Variables: Variables }>();
 
 // Middleware: Set up database connection for all routes
 app.use(async (c, next) => {
@@ -24,21 +28,6 @@ app.use(async (c, next) => {
   const db = drizzle(sql);
   c.set("db", db);
   await next();
-});
-
-// Route Definitions
-// Each route is defined separately with its request/response schema
-// This enables automatic OpenAPI documentation and type safety
-
-const root = createRoute({
-  method: "get",
-  path: "/",
-  responses: {
-    200: {
-      content: { "text/plain": { schema: z.string() } },
-      description: "Root fetched successfully",
-    },
-  },
 });
 
 // Define the expected response shape using Zod
@@ -55,90 +44,53 @@ const UserSchema = z.object({
   email: z.string().email().openapi({
     example: "nikita@neon.tech",
   }),
-}).openapi("User");
+}).openapi({ ref: "User" });
 
-const getUsers = createRoute({
-  method: "get",
-  path: "/api/users",
-  responses: {
-    200: {
-      content: { "application/json": { schema: z.array(UserSchema) } },
-      description: "Users fetched successfully",
-    },
-  },
-});
+const apiRouter = new Hono<{ Bindings: Bindings; Variables: Variables }>();
 
-const NewUserSchema = z.object({
-  name: z.string().openapi({
-    example: "Nikita",
-  }),
-  email: z.string().email().openapi({
-    example: "nikita@neon.tech",
-  }),
-}).openapi("NewUser");
-
-const getUser = createRoute({
-  method: "get",
-  path: "/api/users/{id}",
-  request: {
-    // Validate and parse URL parameters
-    params: z.object({
-      id: z.string().uuid().openapi({
-        example: "123e4567-e89b-12d3-a456-426614174000",
-      }),
-    }),
-  },
-  responses: {
-    200: {
-      content: { "application/json": { schema: UserSchema } },
-      description: "User fetched successfully",
-    },
-  },
-});
-
-const createUser = createRoute({
-  method: "post",
-  path: "/api/users",
-  request: {
-    // Validate request body using Zod schemas
-    body: {
-      required: true, // NOTE: this is important to set to true, otherwise the route will accept empty body
-      content: {
-        "application/json": {
-          schema: NewUserSchema,
+apiRouter.get(
+  "/",
+  describeRoute({
+    responses: {
+      200: {
+        content: {
+          "application/json": { schema: resolver(z.array(UserSchema)) },
         },
+        description: "Users fetched successfully",
       },
     },
-  },
-  responses: {
-    201: {
-      content: {
-        "application/json": {
-          schema: UserSchema,
-        },
-      },
-      description: "User created successfully",
-    },
-  },
-});
-
-// Route Implementations
-// Connect the route definitions to their handlers using .openapi()
-app.openapi(root, async (c) => {
-  return c.text("Honc! 🪿");
-})
-  .openapi(getUsers, async (c) => {
+  }),
+  async (c) => {
     const db = c.get("db");
     const users = await db.select().from(schema.users);
     return c.json(users, 200);
-  })
-  .openapi(getUser, async (c) => {
-    const db = c.get("db");
-    const { id } = c.req.valid("param");
-    const [user] = await db.select().from(schema.users).where(eq(schema.users.id, id));
-    return c.json(user, 200);
-  })
-  .openapi(createUser, async (c) => {
+  },
+).post(
+  "/",
+  describeRoute({
+    responses: {
+      201: {
+        content: {
+          "application/json": {
+            schema: UserSchema,
+          },
+        },
+        description: "User created successfully",
+      },
+    },
+  }),
+  zValidator(
+    "json",
+    z.object({
+      name: z.string().openapi({
+        example: "Nikita",
+      }),
+      email: z.string().email().openapi({
+        example: "nikita@neon.tech",
+      }),
+    }).openapi({ ref: "NewUser" }),
+  ),
+  async (c) => {
     const db = c.get("db");
     const { name, email } = c.req.valid("json");
 
@@ -151,19 +103,71 @@ app.openapi(root, async (c) => {
       .returning();
 
     return c.json(newUser, 201);
-  })
-  // Generate OpenAPI documentation at /openapi.json
-  .doc("/openapi.json", {
-    openapi: "3.0.0",
-    info: {
-      title: "Honc! 🪿",
-      version: "1.0.0",
-      description: "Honc! 🪿",
+  },
+).get(
+  "/:id",
+  describeRoute({
+    responses: {
+      200: {
+        content: { "application/json": { schema: resolver(UserSchema) } },
+        description: "User fetched successfully",
+      },
     },
-  })
-  .use("/fp/*", createFiberplane({
-    app,
-    openapi: { url: "/openapi.json" },
-  }));
+  }),
+  zValidator(
+    "param",
+    z.object({
+      id: z.string().uuid().openapi({
+        example: "123e4567-e89b-12d3-a456-426614174000",
+      }),
+    }),
+  ),
+  async (c) => {
+    const db = c.get("db");
+    const { id } = c.req.valid("param");
+    const [user] = await db.select().from(schema.users).where(
+      eq(schema.users.id, id),
+    );
+    return c.json(user, 200);
+  },
+);
+
+// Route Implementations
+// Connect the route definitions to their handlers using .openapi()
+app.get(
+  "/",
+  describeRoute({
+    responses: {
+      200: {
+        content: { "text/plain": { schema: resolver(z.string()) } },
+        description: "Root fetched successfully",
+      },
+    },
+  }),
+  async (c) => {
+    return c.text("Honc! 🪿");
+  },
+).route("/api/users", apiRouter);
+
+// Generate OpenAPI documentation at /openapi.json
+app.get(
+  "/openapi.json",
+  openAPISpecs(app, {
+    documentation: {
+      info: {
+        title: "Honc! 🪿",
+        version: "1.0.0",
+        description: "Honc! 🪿",
+      },
+    },
+  }),
+)
+  .use(
+    "/fp/*",
+    createFiberplane({
+      app,
+      openapi: { url: "/openapi.json" },
+    }),
+  );
 
 export default app;
