--- templates/d1-hono-openapi/.dev.vars.example
@@ -1,1 +1,0 @@
-DATABASE_URL=
