--- templates/d1-hono-openapi/.prod.vars.example
@@ -1,3 +1,0 @@
-CLOUDFLARE_D1_TOKEN=""
-CLOUDFLARE_ACCOUNT_ID=""
-CLOUDFLARE_DATABASE_ID=""
