--- templates/d1-hono-openapi/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "hono-d1-drizzle-hono-openapi-template",
-  "scripts": {
-    "dev": "wrangler dev src/index.ts",
-    "deploy": "wrangler deploy --minify src/index.ts",
-    "db:touch": "wrangler d1 execute honc-d1-database --local --command='SELECT 1'",
-    "db:generate": "drizzle-kit generate",
-    "db:migrate": "wrangler d1 migrations apply honc-d1-database --local",
-    "db:migrate:prod": "ENVIRONMENT=production drizzle-kit migrate",
-    "db:seed": "tsx seed.ts",
-    "db:seed:prod": "ENVIRONMENT=production tsx seed.ts",
-    "db:setup": "npm run db:touch && npm run db:generate && npm run db:migrate && npm run db:seed",
-    "db:studio": "drizzle-kit studio"
-  },
-  "dependencies": {
-    "@fiberplane/hono": "0.4.4-beta.2",
-    "@hono/zod-validator": "^0.4.3",
-    "dotenv": "^16.4.5",
-    "drizzle-orm": "^0.39.3",
-    "hono": "^4.7.0",
-    "hono-openapi": "^0.4.6",
-    "zod": "^3.23.8",
-    "zod-openapi": "^4.2.4"
-  },
-  "devDependencies": {
-    "@cloudflare/workers-types": "^4.20250321.0",
-    "@libsql/client": "^0.14.0",
-    "drizzle-kit": "^0.30.4",
-    "drizzle-seed": "^0.3.1",
-    "tsx": "^4.19.2",
-    "wrangler": "^4.4.0"
-  }
-}
