--- templates/d1-hono-openapi/wrangler.toml
@@ -1,29 +1,0 @@
-name = "honc-d1-template-openapi"
-compatibility_date = "2025-02-04"
-compatibility_flags = [ "nodejs_compat" ]
-
-[[d1_databases]]
-binding = "DB"
-database_name = "honc-d1-database"
-database_id = "local-honc-d1-database"
-migrations_dir = "drizzle/migrations"
-
-# [vars]
-# MY_VAR = "my-variable"
-
-# Workers Logs
-# Docs: https://developers.cloudflare.com/workers/observability/logs/workers-logs/
-# Configuration: https://developers.cloudflare.com/workers/observability/logs/workers-logs/#enable-workers-logs
-[observability]
-enabled = true
-
-# [[kv_namespaces]]
-# binding = "MY_KV_NAMESPACE"
-# id = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
-
-# [[r2_buckets]]
-# binding = "MY_BUCKET"
-# bucket_name = "my-bucket"
-
-# [ai]
-# binding = "AI"
