+++ templates/d1-openapi/.dev.vars.example
@@ -1,1 +1,1 @@
-ENVIRONMENT="development"
+DATABASE_URL=
