+++ templates/d1-openapi/.prod.vars.example
@@ -1,4 +1,3 @@
-ENVIRONMENT="production"
 CLOUDFLARE_D1_TOKEN=""
 CLOUDFLARE_ACCOUNT_ID=""
 CLOUDFLARE_DATABASE_ID=""
