+++ templates/d1-openapi/seed.ts
@@ -66,7 +66,7 @@ function getLocalD1Db() {
  */
 function getLocalD1DBPath() {
   try {
-    const basePath = path.resolve(".wrangler", "state", "v3", "d1");
+    const basePath = path.resolve(".wrangler");
     const files = fs
       .readdirSync(basePath, { encoding: "utf-8", recursive: true })
       .filter((f) => f.endsWith(".sqlite"));
