--- templates/hono-openapi/seed.ts
@@ -1,29 +1,0 @@
-import { neon } from "@neondatabase/serverless";
-import { config } from "dotenv";
-import { drizzle } from "drizzle-orm/neon-http";
-import * as schema from "./src/db/schema";
-import { seed } from "drizzle-seed";
-
-config({ path: ".dev.vars" });
-
-// biome-ignore lint/style/noNonNullAssertion: error from neon client is helpful enough to fix
-const sql = neon(process.env.DATABASE_URL!);
-const db = drizzle(sql);
-
-async function seedDatabase() {
-  // Read more about seeding here: https://orm.drizzle.team/docs/seed-overview#drizzle-seed
-  await seed(db, schema);
-}
-
-async function main() {
-  try {
-    await seedDatabase();
-    console.log("✅ Database seeded successfully!");
-  } catch (error) {
-    console.error("❌ Error during seeding:", error);
-    process.exit(1);
-  } finally {
-    process.exit(0);
-  }
-}
-main();
